rename tests folder
