+++ tests/auxiliaries.py
