+++ tests/cli_test.py
